refactored test names.
Added queue wait toggle
